Add code to filter `-alltypes` post-conversion build errors believed to reflect known bounds inference limitations. (#24)
Start off the list of filtered errors with "expression has unknown
bounds", as discussed in the PR thread. We will add more soon.

Along with this, stop ignoring non-filtered `-alltypes` post-conversion
build errors. This changes the job status from success to failure. I
think that reflecting reality in this way is helpful (while working on
another task, I had forgotten that the "success" status does not mean
that the builds are successful) and the practical concerns that
originally led us to ignore failure are no longer significant: (1) the
workflow is far from passing as a whole anyway and (2) 3C failures are
now rare, so having the job status indicate 3C success or failure is
less useful.